Fixed loads of /W4 warnings, because it's easier than doing anything useful. Added some asserts to check potentially dangerous assumptions, implemented a few missing bits of code, adjusted some comments, etc.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@814 3db68df2-c116-0410-a063-a993310a9797

+++ source/gui/GUItext.cpp
@@ -240,7 +240,7 @@ void CGUIString::SetValue(const CStr& str)
 		{
 			m_RawString += str.GetSubstring(position, str.Length()-position);
 
-			if (from != m_RawString.Length())
+			if (from != (long)m_RawString.Length())
 			{
 				CurrentTextChunk.m_From = from;
 				CurrentTextChunk.m_To = (int)m_RawString.Length();
